+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 This project follows semantic versioning.
 
+### UNRELEASED (to be 0.4.0)
+
+- [changed] ```Window::new(...)``` now takes WindowOptions struct to configure the creation of the Window.
+- [changed] ```window.update()``` Doesn't take a buffer anymore. See ```window.update_with_buffer```
+- [added] ```window.update_with_buffer()``` Old update version that takes buffer as input parameter
+
 ### v0.3.1 (2016-01-29)
 
 - [fixed] ```get_mouse_pos(MouseMode::Clamp)``` now is in the region [(0, 0) - (width - 1, height - 1)] instead of (width, height)